Get things working, add README
This commit adds a new script to actually replace the function declarations
with the init'd ones. As of this commit, I was able to run all three scripts
to get a successfully build/tested kernel that has 210 new init functions.

I also went ahead an added a README in case others want to give it a shot.