Fix minor bugs
When git grep runs it returns a list of matches in the format:
<call_file>=<call_function>
<call_file>:<regex_match_line>

I had a misunderstanding about this output. I thought that
the above pattern was printed for every regex match. As I now know,
if the same function has the match in two different occassions then
only one <call_file>=<call_function> is printed with multiple <call_file>:<regex_match_line> . This has been fixed in this commit.

Additinally, the function `is_valid_pair` is supposed to ensure this
output format but it was not looking at the regex match line. This has
also been added.

Signed-off-by: Nick Child <nick.child@ibm.com>