Do better parsing when determining if function is __init
To determine if a function is labeled with __init is not as easy and looking
for `__init` in the declaration. Functions like "static int
hash__init_new_context(struct mm_struct *mm)" exist and were causing
functions like `realloc_context_ids` to be added to the group of functions
that could use an __init label. This is false so now we check by determining
if "__init" appears more times in the declaration than in the function name.
This implicitly checks for at least one occurence and it must not be in the
name.

Signed-off-by: Nick Child <nick.child@ibm.com>