Dont investigate functions that are used in macros
Ran into an issue with _dump_db being labeled an
init-able. This is not true since it is assigned to
a macro variable. The macro is then used in other non-
init functions. That being said, other macros like that
for set_pci_dma_ops are okay since it is not assigned
a different name

Signed-off-by: Nick Child <nick.child@ibm.com>